Criado o sistema funcional com a implementação do AstarSearchAlgo
Na linha 39 e 40, é necessário colocar o nome das 2 cidades pra executar
a lógica.

Pra executar o rpgorama pela primeira vez é necessário rodar o metodo
construtor da mapa cotroler, ela ira fornecer o mapa para o banco de
dados.

+++ Eda/bancoDados/model/Cidade.java
@@ -27,7 +27,7 @@ public class Cidade implements Serializable {
     private boolean visitado;
     @ManyToMany
     private List<Adjacente> adjacentes;
-    private int distanciaObjetivo;
+    private double distanciaObjetivo;
 
     /*
     public Cidade(String nome, int distanciaObjetivo) {
@@ -38,7 +38,7 @@ public class Cidade implements Serializable {
     }
     */
     
-    public int getDistanciaObjetivo() {
+    public double getDistanciaObjetivo() {
 		return distanciaObjetivo;
 	}
